backend/pyproject.toml
[project]
name = "lexdeepresearch"
version = "0.1.0"
description = "自动化信息获取与深度研究平台"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.0.0",
    "apscheduler>=3.10.0",
    "litellm>=1.50.0",
    "feedparser>=6.0.0",
    "httpx>=0.28.0",
    "selectolax>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.52.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
